requirements.txt
requests
beautifulsoup4
pandas
openpyxl
snowflake-connector-python
python-dotenv
lxml